Refactor imports to streamline structure access and update version to v0.1.7
