+++ kmcpy/external/structure.py
@@ -26,12 +26,12 @@ class StructureKMCpy(Structure):
             merge_tol (float, optional): not implemented?. Defaults to 0.0.
 
         Returns:
-            kmcpy.external.pymatgen_structure.Istructure: a structure generated by the cif with all information you need.
+            kmcpy.external.StructureKMCpy: a structure generated by the cif with all information you need.
         """
         # modified version of from file, only read the cif
         filename = str(filename)
 
-        from kmcpy.external.cif import CifParserKMCpy
+        from kmcpy.external import CifParserKMCpy
 
         fname = os.path.basename(filename)
         with zopen(filename, "rt") as f:
@@ -111,7 +111,7 @@ class StructureKMCpy(Structure):
         """
         if "wyckoff_sequence" not in self.site_properties:
             raise ValueError(
-                "wyckoff_sequence not found in site properties.make sure that this structure is generated by kmcpy.external.pymatgen_structure.from_cif() function!"
+                "wyckoff_sequence not found in site properties.make sure that this structure is generated by kmcpy.StructureKMCpy.from_cif() function!"
             )
         for i in self._sites:
             if (
@@ -150,7 +150,7 @@ class StructureKMCpy(Structure):
         """
         if "wyckoff_sequence" not in self.site_properties:
             raise ValueError(
-                "wyckoff_sequence not found in site properties.make sure that this structure is generated by kmcpy.external.pymatgen_structure.from_cif() function!"
+                "wyckoff_sequence not found in site properties.make sure that this structure is generated by kmcpy.external.StructureKMCpy.from_cif() function!"
             )
         if return_index:
             for i in range(0, len(self._sites)):
@@ -186,7 +186,7 @@ class StructureKMCpy(Structure):
             scaling_matrix (tuple, optional): scaling matrix . Defaults to (1,2,3).
 
         Returns:
-            Structure: kmcpy.external.pymatgen_structure.Structure
+            Structure: kmcpy.external.StructureKMCpy
         """
 
         scale_matrix = np.array(scaling_matrix, int)
